+++ lib/data/fusion_data.dart
@@ -1,0 +1,4 @@
+class FusionData {
+  static late Duration transitionDuration;
+  static late Duration reverseTransitionDuration;
+}